python-workers/requirements.txt
kafka-python
transformers
torch
